Add example page, fix RawResource.Limit type, and register service as scoped

+++ src/Web/Models/RawResource.cs
@@ -8,5 +8,5 @@ public class RawResource
     public string Name { get; set; } = string.Empty;
 
     [JsonPropertyName("limit")]
-    public int Limit { get; set; }
+    public long Limit { get; set; }
 }
